maven enabled, implemenation changed to support hbase-0.89.x dependency
