+++ .gitignore
@@ -1,1 +1,4 @@
-build/
+.project
+.classpath
+.settings
+target
