+++ src/main/resources/schema.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configurations>
+	<configuration>
+		<name>local</name>
+		<description>Configuration for the HBase cluster.</description>
+		<hbase_master>localhost:60000</hbase_master>
+		<zookeeper_quorum>localhost</zookeeper_quorum>
+		<zookeeper_client_port>2181</zookeeper_client_port>
+		<schema>
+			<table>
+				<name>tablename</name>
+				<description>A table</description>
+				<column_family>
+					<name>columnA</name>
+					<description>Column A.</description>
+					<!-- Default: 3 -->
+					<max_versions />
+					<!-- Default: None -->
+					<compression />
+					<!-- Default: false -->
+					<in_memory />
+					<!-- Default: true -->
+					<block_cache_enabled />
+					<!-- Default: 64k -->
+					<block_size />
+					<!-- Default: Integer.MAX_VALUE (forever) -->
+					<time_to_live />
+					<!-- Default: false -->
+					<bloom_filter />
+					<scope />
+					<!-- Default: 0-->
+				</column_family>
+				<column_family>
+					<name>columnB</name>
+					<description>Column B.</description>
+					<!-- Default: 3 -->
+					<max_versions />
+					<!-- Default: None -->
+					<compression />
+					<!-- Default: false -->
+					<in_memory />
+					<!-- Default: true -->
+					<block_cache_enabled />
+					<!-- Default: 64k -->
+					<block_size />
+					<!-- Default: Integer.MAX_VALUE (forever) -->
+					<time_to_live />
+					<!-- Default: false -->
+					<bloom_filter />
+					<scope />
+					<!-- Default: 0-->
+				</column_family>
+			</table>
+		</schema>
+	</configuration>
+</configurations>
